% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_prism.R
\name{model_run}
\alias{model_run}
\title{Run CHDWILSON Model}
\usage{
model_run(model_input = NULL)
}
\arguments{
\item{model_input}{A list/json object with gender, age,
TChol, LDL,HDL,SBP,DBP,diabetes and smoker}
}
\value{
Return a dataset with two columns: CHDprediction_use_TChol, CHDprediction_use_ldl
}
\description{
This function calls the CHDWILSON model.
}
